requirements.txt
PySide6>=6.6.0
PySide6-Addons>=6.6.0
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
scikit-image>=0.21.0
pyinstaller>=6.0.0
